requirements.txt
fastapi
uvicorn
gunicorn
pydantic
python-dotenv
openai
redis
edge-tts
langdetect
sentence-transformers
langchain-text-splitters
pymongo
dnspython